@@ -1,22 +1,21 @@
 ﻿using System;
 using System.Net;
 
-namespace AutoUpdaterDotNET
-{
-    /// <inheritdoc />
-    public class MyWebClient : WebClient
-    {
-        /// <summary>
-        ///     Response Uri after any redirects.
-        /// </summary>
-        public Uri ResponseUri;
+namespace AutoUpdaterDotNET;
 
-        /// <inheritdoc />
-        protected override WebResponse GetWebResponse(WebRequest request, IAsyncResult result)
-        {
-            WebResponse webResponse = base.GetWebResponse(request, result);
-            ResponseUri = webResponse.ResponseUri;
-            return webResponse;
-        }
+/// <inheritdoc />
+public class MyWebClient : WebClient
+{
+    /// <summary>
+    ///     Response Uri after any redirects.
+    /// </summary>
+    public Uri ResponseUri;
+
+    /// <inheritdoc />
+    protected override WebResponse GetWebResponse(WebRequest request, IAsyncResult result)
+    {
+        WebResponse webResponse = base.GetWebResponse(request, result);
+        ResponseUri = webResponse.ResponseUri;
+        return webResponse;
     }
 }
