@@ -27,7 +27,8 @@ namespace ZipExtractor
         {
             string zipPath = null;
             string extractionPath = null;
-            string executablePath = null;
+            string currentExe = null;
+            string updatedExe = null;
             bool clearAppDirectory = false;
             string commandLineArgs = null;
 
@@ -47,8 +48,11 @@ namespace ZipExtractor
                     case "--output":
                         extractionPath = args[index + 1];
                         break;
-                    case "--executable":
-                        executablePath = args[index + 1];
+                    case "--current-exe":
+                        currentExe = args[index + 1];
+                        break;
+                    case "--updated-exe":
+                        updatedExe = args[index + 1];
                         break;
                     case "--clear":
                         clearAppDirectory = true;
@@ -69,7 +73,7 @@ namespace ZipExtractor
 
             _logBuilder.AppendLine();
 
-            if (string.IsNullOrEmpty(zipPath) || string.IsNullOrEmpty(extractionPath) || string.IsNullOrEmpty(executablePath))
+            if (string.IsNullOrEmpty(zipPath) || string.IsNullOrEmpty(extractionPath) || string.IsNullOrEmpty(currentExe))
             {
                 return;
             }
@@ -83,11 +87,11 @@ namespace ZipExtractor
 
             _backgroundWorker.DoWork += (_, eventArgs) =>
             {
-                foreach (var process in Process.GetProcessesByName(Path.GetFileNameWithoutExtension(executablePath)))
+                foreach (var process in Process.GetProcessesByName(Path.GetFileNameWithoutExtension(currentExe)))
                 {
                     try
                     {
-                        if (process.MainModule is { FileName: { } } && process.MainModule.FileName.Equals(executablePath))
+                        if (process.MainModule is { FileName: { } } && process.MainModule.FileName.Equals(currentExe))
                         {
                             _logBuilder.AppendLine("Waiting for application process to exit...");
 
@@ -103,11 +107,11 @@ namespace ZipExtractor
 
                 _logBuilder.AppendLine("BackgroundWorker started successfully.");
 
-                    // Ensures that the last character on the extraction path
-                    // is the directory separator char.
-                    // Without this, a malicious zip file could try to traverse outside of the expected
-                    // extraction path.
-                    if (!extractionPath.EndsWith(Path.DirectorySeparatorChar.ToString(), StringComparison.Ordinal))
+                // Ensures that the last character on the extraction path
+                // is the directory separator char.
+                // Without this, a malicious zip file could try to traverse outside of the expected
+                // extraction path.
+                if (!extractionPath.EndsWith(Path.DirectorySeparatorChar.ToString(), StringComparison.Ordinal))
                 {
                     extractionPath += Path.DirectorySeparatorChar;
                 }
@@ -163,9 +167,17 @@ namespace ZipExtractor
                                 if (!entry.IsDirectory())
                                 {
                                     var parentDirectory = Path.GetDirectoryName(filePath);
-                                    if (!Directory.Exists(parentDirectory))
+
+                                    if (parentDirectory != null)
                                     {
-                                        Directory.CreateDirectory(parentDirectory);
+                                        if (!Directory.Exists(parentDirectory))
+                                        {
+                                            Directory.CreateDirectory(parentDirectory);
+                                        }
+                                    }
+                                    else
+                                    {
+                                        throw new ArgumentNullException($"parentDirectory is null for \"{filePath}\"!");
                                     }
                                     using (Stream destination = File.Open(filePath, FileMode.OpenOrCreate, FileAccess.Write, FileShare.None))
                                     {
@@ -183,50 +195,48 @@ namespace ZipExtractor
                                 const int errorSharingViolation = 0x20;
                                 const int errorLockViolation = 0x21;
                                 var errorCode = Marshal.GetHRForException(exception) & 0x0000FFFF;
-                                if (errorCode is errorSharingViolation or errorLockViolation)
+                                if (errorCode is not (errorSharingViolation or errorLockViolation))
                                 {
-                                    retries++;
-                                    if (retries > MaxRetries)
-                                    {
-                                        throw;
-                                    }
+                                    throw;
+                                }
 
-                                    List<Process> lockingProcesses = null;
-                                    if (Environment.OSVersion.Version.Major >= 6 && retries >= 2)
-                                    {
-                                        try
-                                        {
-                                            lockingProcesses = FileUtil.WhoIsLocking(filePath);
-                                        }
-                                        catch (Exception)
-                                        {
-                                                // ignored
-                                            }
-                                    }
+                                retries++;
+                                if (retries > MaxRetries)
+                                {
+                                    throw;
+                                }
 
-                                    if (lockingProcesses == null)
+                                List<Process> lockingProcesses = null;
+                                if (Environment.OSVersion.Version.Major >= 6 && retries >= 2)
+                                {
+                                    try
                                     {
-                                        Thread.Sleep(5000);
+                                        lockingProcesses = FileUtil.WhoIsLocking(filePath);
                                     }
-                                    else
+                                    catch (Exception)
                                     {
-                                        foreach (var lockingProcess in lockingProcesses)
-                                        {
-                                            var dialogResult = MessageBox.Show(
+                                        // ignored
+                                    }
+                                }
+
+                                if (lockingProcesses == null)
+                                {
+                                    Thread.Sleep(5000);
+                                }
+                                else
+                                {
+                                    foreach (var lockingProcess in lockingProcesses)
+                                    {
+                                        var dialogResult = MessageBox.Show(
                                                 string.Format(Resources.FileStillInUseMessage,
                                                     lockingProcess.ProcessName, filePath),
                                                 Resources.FileStillInUseCaption,
                                                 MessageBoxButtons.RetryCancel, MessageBoxIcon.Error);
-                                            if (dialogResult == DialogResult.Cancel)
-                                            {
-                                                throw;
-                                            }
+                                        if (dialogResult == DialogResult.Cancel)
+                                        {
+                                            throw;
                                         }
                                     }
-                                }
-                                else
-                                {
-                                    throw;
                                 }
                             }
                         }
@@ -263,27 +273,27 @@ namespace ZipExtractor
                         throw eventArgs.Error;
                     }
 
-                    if (!eventArgs.Cancelled)
+                    if (eventArgs.Cancelled) return;
+
+                    textBoxInformation.Text = @"Finished";
+                    try
                     {
-                        textBoxInformation.Text = @"Finished";
-                        try
+                        var executablePath = string.IsNullOrWhiteSpace(updatedExe) ? currentExe : Path.Combine(extractionPath, updatedExe);
+                        ProcessStartInfo processStartInfo = new ProcessStartInfo(executablePath);
+                        if (!string.IsNullOrEmpty(commandLineArgs))
                         {
-                            ProcessStartInfo processStartInfo = new ProcessStartInfo(executablePath);
-                            if (!string.IsNullOrEmpty(commandLineArgs))
-                            {
-                                processStartInfo.Arguments = commandLineArgs;
-                            }
-
-                            Process.Start(processStartInfo);
-
-                            _logBuilder.AppendLine("Successfully launched the updated application.");
+                            processStartInfo.Arguments = commandLineArgs;
                         }
-                        catch (Win32Exception exception)
+
+                        Process.Start(processStartInfo);
+
+                        _logBuilder.AppendLine("Successfully launched the updated application.");
+                    }
+                    catch (Win32Exception exception)
+                    {
+                        if (exception.NativeErrorCode != 1223)
                         {
-                            if (exception.NativeErrorCode != 1223)
-                            {
-                                throw;
-                            }
+                            throw;
                         }
                     }
                 }
